+++ src/app/styles/common/theme.css
@@ -310,7 +310,7 @@
   --palette-gray-70: #404040;
   --palette-gray-80: #262626;
   --palette-gray-90: #0d0d0d;
-
+  --palette-gray-100: #f4f4f4;
   /* White */
   --palette-white-05: #ffffff;
   --palette-white-10: #fafafa;
