+++ src/features/auth/signUp/steps/signupsteptype.ts
@@ -1,0 +1,4 @@
+export type SignUpStepProps = {
+  onPrev?: () => void;
+  onNext?: () => void;
+};
